Add files via upload
Added Functionality of aerofoil input, as well as specifying how many points to be left in the trailing edge truncation. Possibly useful for turbulent wakes or something. Could well be smoothed out towards the boundary. The smoothing is to be added. Will attempt to add Laplace-Beltrami smoothing, as this appears to give very good smoothing, and I have found a fully laid out implementation